tee-supplicant: move file operations and flags to their own header
The flags and conversion functions will be re-used by the SQL
filesystem, so move them to a common header: tee_fs.h.
